split Position classes out to their own files
