+++ lib/position/filter_html.rb
