+++ rake_build/generate_final_csvs.rb
@@ -1,4 +1,3 @@
-require_relative '../lib/position_filterer'
 require 'everypolitician/popolo'
 require 'json5'
 
@@ -111,118 +110,6 @@ namespace :term_csvs do
     matched, unmatched = wikidata_parties.map(&:count)
     warn "Parties matched to Wikidata: #{matched} ✓ #{unmatched.zero? ? '' : "| #{unmatched} ✘"}"
     wikidata_parties.last.shuffle.take(5).each { |p| warn "  No wikidata: #{p[:name]} (#{p[:id]})" } unless matched.zero?
-  end
-
-  # TODO: move this to its own file
-  class PositionMap
-    # Which Wikidata Positions we're interested in, and how to group them
-
-    def initialize(pathname:)
-      @pathname = pathname
-    end
-
-    def to_json
-      return empty_filter unless pathname.exist?
-      raw_json
-    end
-
-    def include_ids
-      to_json[:include].values.flatten.map { |p| p[:id] }.to_set
-    end
-
-    def exclude_ids
-      to_json[:exclude].values.flatten.map { |p| p[:id] }.to_set
-    end
-
-    def known_ids
-      include_ids + exclude_ids
-    end
-
-    def cabinet_ids
-      (to_json[:include][:cabinet] || []).map { |p| p[:id] }.to_set
-    end
-
-    private
-
-    attr_reader :pathname
-
-    def empty_filter
-      { exclude: { self: [], other: [] }, include: { self: [], other_legislatures: [], cabinet: [], executive: [], party: [], other: [] } }
-    end
-
-    def raw_json
-      @json ||= json5_parse(pathname.read).each do |_s, fs|
-        fs.each { |_, fs| fs.each { |f| f.delete :count } }
-      end
-    end
-
-    # TODO: move this to somewhere more generally useful
-    def json5_parse(data)
-      # read with JSON5 to be more liberal about trailing commas.
-      # But that doesn't have a 'symbolize_names' so rountrip through JSON
-      JSON.parse(JSON5.parse(data).to_json, symbolize_names: true)
-    end
-  end
-
-  class WikidataPositionFile
-    def initialize(pathname:)
-      @pathname = pathname
-    end
-
-    def positions_for(person)
-      json[person.wikidata.to_sym].to_a.map do |posn|
-        WikidataPosition.new(raw: posn, person: person)
-      end
-    end
-
-    def json
-      JSON.parse(pathname.read, symbolize_names: true)
-    end
-
-    private
-
-    attr_reader :pathname
-  end
-
-
-  class WikidataPosition
-    attr_reader :person
-    def initialize(raw:, person:)
-      @raw = raw
-      @person = person
-    end
-
-    def id
-      raw[:id]
-    end
-
-    def label
-      raw[:label]
-    end
-
-    def description
-      raw[:description]
-    end
-
-    def start_date
-      qualifier(580)
-    end
-
-    def end_date
-      qualifier(582)
-    end
-
-    private
-
-    attr_reader :raw
-
-    def qualifiers
-      raw[:qualifiers] || {}
-    end
-
-    def qualifier(pcode)
-      qualifiers["P#{pcode}".to_sym]
-    end
   end
